+++ lib/src/stub/permission_manager_stub.dart
@@ -1,4 +1,5 @@
 // PermissionManager stub
+import 'package:flutter/material.dart';
 import 'package:adaptive_media_picker/adaptive_media_picker.dart';
 
 class PermissionManager {
@@ -9,6 +10,7 @@ class PermissionManager {
   Future<PermissionResolution> ensureMediaPermission({
     required ImageSource source,
     required MediaType mediaType,
+    BuildContext? context,
   }) async {
     return PermissionResolution.grantedFull();
   }